requirements.txt
duckdb>=1.1.0
pyyaml>=6.0
minio>=7.2
pandas>=2.0
